Add PHPCompatibility standards and install scripts
The PHPCompatibility standards is a PHP_CodeSniffer standard designed to check whether PHP code is compatible with different versions of PHP.

Here, we include the sniff, and set it (and the WPCS standards) up by default when composer is updated.

Currently pinned against `dev-master`, as that contains a fix since 7.1.4, and 7.1.5 is not released yet.